requirements.txt
Django>=5.2.8
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
psycopg2-binary
Pillow
reportlab